refactor quantize, finalize atomic quantize
