+++ python/hidet/graph/ops/quant/__init__.py
@@ -10,4 +10,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from .symmetric import symmetric_quantize, symmetric_dequantize
-from .matmul import symmetric_quant_matmul, symmetric_quant_matmulf16
+from .matmul import symmetric_quant_matmul
+from .matmul_f16_i8 import symmetric_quant_matmul_f16_i8
+from .matmul_f16_i8_atomic import symmetric_quant_matmul_atomic_f16_i8
